Import dialog layout cleaned
Buttons were messed up because of the inheritance from KDialog.
Removed UI file importdialog.ui, changed in a widget -> importwidget.ui
This widget is the main widget class of the dialog, buttons are set
to KDE default (more integration with the rest of KDE) and renamed
as it was before in case of editing.

Signed-off-by: Riccardo Bellini <ricky88ykcir@gmail.com>
Signed-off-by: Brian Korbein <bri.kor.21@gmail.com>