Add autocompletion features for authors and genres
Added a QueryEngine class, an interface for the actual queries
execution. This class is called from ImportDialog.
QueryEngine has an helper class, QueryEngineInternal, which performs
the actual query. The query is executed in a separate thread, to
keep UI reponsive. Data are moved from the query to the import dialog
using convenience signals from QueryEngine class.

Signed-off-by: Riccardo Bellini <ricky88ykcir@gmail.com>