Create a dialog to modify one or more books.
Class ModifyDialog allows the user to edit data of one or more than
one book. Special  "Previous" and "Next" button are provided,
to navigate through the entries.
There is also a checkbox which allows the user to edit multiple kind
of data at the same time, for all selected books.

Signed-off-by: Riccardo Bellini <ricky88ykcir@gmail.com>