Modified buildIcons structure
Now the IconBuilder worker thread takes in input a list of books
and emits a signal for each fetched and cached preview.
PreviewRole of the item is updated to store the key of the image
in the cache.
TODO: logic still incomplete, need of a rework in tree model's data
structure.